+++ src/main/javascript/command/dpat-server.js
@@ -5,6 +5,7 @@ const project = require("../Project").newInstance();
 const __path = require("path");
 const defaultWebpackConfig = require.resolve('../webpack/webpack.config-development');
 const webpackDevServerPath = __path.resolve(__dirname, '../../../../node_modules/.bin/webpack-dev-server');
+const dpatModules = __path.resolve(__dirname, '../../../../node_modules');
 /**
  * @param {String} path
  * @param {Command} cmd
@@ -13,7 +14,12 @@ function action(path, port, cmd)
 {
     const projectDir = __path.resolve(path);
 
-    if (! project.startDevServer(projectDir, webpackDevServerPath, defaultWebpackConfig)) {
+    const runConfig = {
+        webpackDev: webpackDevServerPath,
+        webpackConfig: defaultWebpackConfig,
+        modulePaths: [ dpatModules ]
+    };
+    if (! project.startDevServer(projectDir, runConfig)) {
         console.error(`ERROR: could not start dev server at ${path}`);
         process.exit(1);
     }